requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
